Update to fix 500 No key

+++ src/main/java/com/example/iotimList/repos/TaskRepos.java
@@ -3,5 +3,5 @@ package com.example.iotimList.repos;
 import com.example.iotimList.domain.Task;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TaskRepos extends CrudRepository<Task, Integer> {
+public interface TaskRepos extends CrudRepository<Task, Long> {
 }
